requirements.txt
pandas>=2.0
psycopg2-binary>=2.9
httpx>=0.25
supabase>=2.0
python-dotenv>=1.0
tqdm>=4.66
requests>=2.31
